requirements/cicd_utils.txt
# ./cicd_utils/cicd/compile_plotly_charts.py
minify-html
kaleido

# ./cicd_utils/cicd/scripts/extract_latest_release_notes.py
markdown-it-py
mdformat
